WIP: improving community

+++ src/html/scripts/community.js
@@ -1,8 +1,10 @@
 var user = null;
 var pageMembers = 0;
+var edit = false;
 
 
 // TODO: divide createcommunity function in separate module and fix
+
 
 async function authAndSetUser(endpoint) {
    try {
@@ -18,13 +20,13 @@ async function authAndSetUser(endpoint) {
          throw new Error("Utente non loggato");
       } else {
          user = await userData;
-         console.log(user);
          checkIfUserHasCommunity(endpoint, user._id);
       }
    } catch (error) {
       console.error("Errore durante l'esecuzione:", error);
    }
 }
+
 
 async function checkIfUserHasCommunity(endpoint, uid) {
    // TODO: improve if user is not logged
@@ -33,7 +35,7 @@ async function checkIfUserHasCommunity(endpoint, uid) {
       return;
    }
    let community = await fetchCommunity(uid);
-   console.log(endpoint, uid, community);
+   // console.log(endpoint, uid, community);
 
    switch (endpoint) {
       case 'createcommunity':
@@ -70,53 +72,15 @@ async function checkIfUserHasCommunity(endpoint, uid) {
          };
 
          populateMembers(community.members, endpoint);
-         communityHandleUI(uid, community);
+         getCommunityPlaylists(community.playlists, endpoint);
+         // populatePlaylists(community.playlists, endpoint);
+         communityHandleUI(community);
          break;
 
       default: break;
    }
 }
 
-
-// async function fetchUsersAndUpdateCards(membersData) {
-//    // TODO: needs button to add users to community
-//    if (membersData.lenght == null) return;
-//    try {
-//       membersData.forEach(async member => {
-//          let res = await fetch(`/users/${member}`);
-//          let memberInfo = await res.json()
-//          // console.log(memberInfo, memberInfo[0].nickname)
-//          const membersContainer = document.getElementById('membersContainer');
-//          const card = `
-// <div class="col-md-4 mb-4">
-// <div class="card">
-// <div class="card-body">
-// <h5 class="card-title">${memberInfo[0].nickname}</h5>
-// </div>
-// </div>
-// </div>
-// `;
-//          membersContainer.innerHTML += card;
-//          // return memberInfo
-//       });
-
-//    } catch (error) {
-//       console.error("Errore durante il recupero degli utenti:", error);
-//       return {};
-//    }
-// }
-
-
-// function editCommunity(community) {
-//    console.log("Editing community...");
-//    /**
-//     * local updateCommunity = await fetch(`/community/`, {
-//     * method: 'PUT'
-//     * body: ''
-//     * body: ''
-//     * })
-//     */
-// }
 
 
 // function prevMembers() {
@@ -132,16 +96,27 @@ async function checkIfUserHasCommunity(endpoint, uid) {
 //    getCredits(id, pageResults)
 // }
 
+
 function getUserPlaylists(uid, endpoint) {
    fetch(`/playlist/${uid}`).then((response) => {
       if (response.ok) {
-         response.json().then((playlists) => {
-            populatePlaylists(playlists, endpoint);
+         response.json().then((userPlaylists) => {
+            populatePlaylists(userPlaylists, endpoint);
          })
       }
-   })
+   });
 }
 
+async function getCommunityPlaylists(playlists, endpoint) {
+   let playlistsData = [];
+
+   for await (const i of playlists) {
+      let res = await fetch(`/getplaylist/${i.pid}`);
+      let pl = await res.json();
+      playlistsData.push(pl);
+   }
+   populatePlaylists(playlistsData, endpoint);
+}
 
 
 function populateMembers(members, endpoint) {
@@ -191,10 +166,10 @@ function populateMembers(members, endpoint) {
       case 'community':
          // TODO: add button to remove member from community
 
-         for (const member in members) {
+         for (let i in members) {
             /* Skip card creation for creator of community */
-            if (user._id == member._id) continue;
-            fetch(`/users/${members[member]._id}`).then((response) => {
+            if (user._id == members[i].uid) continue;
+            fetch(`/users/${members[i].uid}`).then((response) => {
                if (response.ok) {
                   response.json().then((m) => {
                      var clone = card.cloneNode(true)
@@ -227,6 +202,7 @@ function populateMembers(members, endpoint) {
    }
 }
 
+
 function populatePlaylists(playlists, endpoint) {
    var card = document.getElementById('card-cast');
    var container = document.getElementById('container-cast');
@@ -235,7 +211,7 @@ function populatePlaylists(playlists, endpoint) {
    // console.log(playlists, endpoint)
    switch (endpoint) {
       case 'createcommunity':
-         const playlistContainer = document.getElementById("playlistContainer");
+         let playlistContainer = document.getElementById("playlistContainer");
          playlists.forEach(playlist => {
             var stringified = JSON.stringify(playlist.songs).replace(/"/g, '&quot;');
             const card = `
@@ -259,43 +235,67 @@ function populatePlaylists(playlists, endpoint) {
          break;
 
       case 'community':
-         // TODO: add button to remove member from community
+         let playlistPublicContainer = document.getElementById("playlistPublicContainer");
 
-         for (const member in members) {
-            /* Skip card creation for creator of community */
-            if (user._id == members[i]._id) continue;
-            fetch(`/users/${members[member]._id}`).then((response) => {
-               if (response.ok) {
-                  response.json().then((m) => {
-                     var clone = card.cloneNode(true)
-                     // console.log(m);
-                     clone.id = 'card-cast-' + m._id
-                     clone.getElementsByClassName('card-text')[0].innerHTML = m.name
-                     clone.getElementsByClassName('text-body-secondary')[0].innerHTML = m.nickname
+         // const creatorAddPlaylist = `
+         // <div class="col-md-4 mb-4">
+         //   <div class="card h-100">
+         //    <div class="card-body">
+         //     <button class="btn btn-danger" onclick="creatorAddPlaylist('${user._id}')" hidden disabled>
+         //         Remove
+         //     </button>
+         //    </div>
+         //   </div>
+         // </div>`;
+         // playlistPublicContainer += creatorAddPlaylist;
 
-                     // IDT we want to add the profile picture for users... just use a placeholder
-                     clone.getElementsByClassName('card-img-top')[0].src =
-                        'https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png'
+         // console.log(playlistPublicContainer);
+         playlists.forEach(playlist => {
+           var stringified = JSON.stringify(playlist.songs).replace(/"/g, '&quot;');
+           const card = `
+           <div class="col-md-4 mb-4">
+           <div class="card h-100">
+               <div class="card-body">
+                   <h5 class="card-title">${playlist.title}${playlist.private ? '<i class="private bi bi-lock-fill text-success"></i>' : ''}</h5>
+                   <p class="card-text">${playlist.description}</p>
+                   <button class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#songsModal${playlist._id}"
+                       onclick="showSongs('${playlist._id}', ${stringified})">
+                       View Songs
+                   </button>
+                   <button class="btn btn-danger" onclick="removePlaylist('${playlist._id}','${playlist.title}')" hidden disabled>
+                       Remove
+                   </button>
 
-                     let strMember = JSON.stringify(m);
-                     clone.getElementsByClassName('btn btn-danger')[0]
-                        .setAttribute('onClick', `removeMemberFromCommunity('${strMember}')`);
-
-                     clone.classList.remove('d-none')
-                     card.before(clone)
-                  })
-               }
-            }).catch((err) => {
-                  console.error("Error fetching members of community:", err);
-                  alert("Error fetching members of community. Retry!");
-               })
-         }
-
+               </div>
+           </div>
+          </div>`;
+           playlistPublicContainer.innerHTML += card;
+        });
          break;
 
       default: break;
    }
 }
+
+
+function removePlaylist(pid, title) {
+   if (!window.confirm(`Confirm remove playlist < ${title} > from community?`)) return;
+
+   fetch(`/community/${user._id}`, {
+      method: 'PUT',
+      headers: {
+         'Content-Type': 'application/json;charset=utf-8'
+      },
+      body: JSON.stringify({ op: 'removePlaylist', creatorId: user._id, pid: pid })
+   }).then(response => {
+         console.log(response.json());
+         // alert(response);
+         if (response.ok) {
+            window.location.replace('/community');
+         }
+      }).catch(err => console.error(err));
+}
+
 
 function addUserToCommunityMembers(userId) {
    let checkbox = document.getElementById('user-selection-'+userId);
@@ -324,9 +324,11 @@ function addUserToCommunityMembers(userId) {
       });
 }
 
+
 function removeMemberFromCommunity(member) {
    let strObj = JSON.parse(member);
-   console.log("To remove from community: ", strObj);
+   if (!window.confirm(`Confirm remove < ${strObj.nickname} > from community?`)) return;
+
    fetch(`/community/${user._id}`, {
       method: 'PUT',
       headers: {
@@ -335,7 +337,6 @@ function removeMemberFromCommunity(member) {
       body: JSON.stringify({ op: 'removeMember', creatorId: user._id, member: member })
    }).then(response => {
          console.log(response.json());
-         alert(response);
          if (response.ok) {
             window.location.replace('/community');
          }
@@ -350,49 +351,69 @@ async function fetchCommunity(creatorId) {
 }
 
 
-function communityHandleUI(uid, community) {
+function communityHandleUI(community) {
    const communityContainer = document.getElementById('community-container');
    const communityGeneralButton = document.getElementById("communityGeneralButton");
-   const deleteCommunityButton = document.getElementById("deleteCommunityButton");
 
    document.getElementById('community-title').innerText = community.name;
    document.getElementById('community-desc').innerText = community.desc;
 
    // if (community.creatorId == user._id) {
    communityGeneralButton.removeAttribute('disabled');
-   deleteCommunityButton.removeAttribute('disabled');
 
    communityGeneralButton.addEventListener("click", () => {
-      // TODO: manage edit of community:
-      //        - add/remove members
-      //        - add/remove playlist
-      editCommunity(community);
-      // window.location.href = '/createcommunity';
+      toggleEditCommunity(community);
    });
+}
 
 
-   deleteCommunityButton.addEventListener("click", () => {
-      // console.log(user);
-      if (window.confirm("Do you really want to delete this community?")) {
-         try {
-            fetch(`/community/${user.id}`, {
-               method: 'DELETE',
-               headers: {
-                  'Content-Type': 'application/json;charset=utf-8'
-               },
-               body: JSON.stringify({ creatorId: user._id})
-            }).then(response => {
-                  console.log(response.json());
-                  alert(response);
-                  if (response.ok) {
-                     window.location.replace('/community');
-                  }
-               })
-         } catch (err) {
-            console.error(err)
+function toggleEditCommunity(community) {
+   if (edit) {
+      edit = false;
+      let btns = document.querySelectorAll('.btn-danger');
+      btns.forEach(btn => {
+         btn.setAttribute('disabled', true);
+         btn.setAttribute('hidden', true);
+      })
+
+   } else {
+      edit = true;
+      let btns = document.querySelectorAll('.btn-danger');
+      btns.forEach(btn => {
+         btn.removeAttribute('disabled');
+         btn.removeAttribute('hidden');
+      })
+
+      const deleteCommunityButton = document.getElementById("deleteCommunityButton");
+      deleteCommunityButton.addEventListener("click", () => {
+         // console.log(user);
+         if (window.confirm("Do you really want to delete this community?")) {
+            try {
+               fetch(`/community/${user.id}`, {
+                  method: 'DELETE',
+                  headers: {
+                     'Content-Type': 'application/json;charset=utf-8'
+                  },
+                  body: JSON.stringify({ creatorId: user._id})
+               }).then(response => {
+                     console.log(response.json());
+                     alert(response);
+                     if (response.ok) {
+                        window.location.replace('/community');
+                     }
+                  })
+            } catch (err) {
+               console.error(err)
+            }
          }
-      }
-   });
+      });
+
+   }
+
+
+   // deleteCommunityButton.removeAttribute('disabled');
+   // deleteCommunityButton.removeAttribute('hidden');
+
 }
 
 
@@ -406,8 +427,8 @@ function checkFieldFullfilled() {
 }
 
 
+
 function searchUserToAdd() {
-   // TODO: add selection to users not present already in community
    fetch('/users').then((response) => {
       if (response.ok) {
          response.json().then(data => console.log(data))
@@ -415,8 +436,3 @@ function searchUserToAdd() {
    }).catch((err) => console.error(err));
 }
 
-
-// TODO: to be done!
-function getPlaylistsOfCommunity(params) {
-
-}
